modify(point): 포인트 이벤트 리스너 통합
